pyproject.toml
[project]
name = "ulcerative-colitis"
version = "0.1.0"
authors = [
    {name = "Adam Kukučka", email = "524905@muni.cz"},
]
requires-python = ">=3.11"
readme = "README.md"
license = { file = "LICENSE" }
dependencies = [
    "torch>=2.4.0",
    "lightning>=2.4.0",
    "torchmetrics>=1.4.1",
    "torchvision>=0.17.2",
    "albumentations>=1.4.14",
    "ipykernel",
    "rationai-mlkit @ git+https://gitlab.ics.muni.cz/rationai/digital-pathology/libraries/mlkit.git@fix/metrics",
    "rationai-masks @ git+https://gitlab.ics.muni.cz/rationai/digital-pathology/libraries/masks.git",
    "rationai-tiling @ git+https://gitlab.ics.muni.cz/rationai/digital-pathology/libraries/tiling.git",
]

[tool.pdm.dev-dependencies]
dev = ["mypy", "pre-commit", "ruff"]

[tool.pdm.scripts]
train_neutrophils = "python -m ulcerative_colitis --config-name=train_neutrophils"
test_neutrophils = "python -m ulcerative_colitis --config-name=test_neutrophils"
predict_neutrophils = "python -m ulcerative_colitis --config-name=predict_neutrophils"
l = { composite = ["lint", "format", "mypy"] }
lint = "ruff check --fix"
format = "ruff format"
mypy = "mypy ."
post_install = { composite = [
    "pre-commit autoupdate",
    "pre-commit install",
    "pre-commit install --hook-type commit-msg",
] }

[tool.pdm]
distribution = false
